token

+++ asgn3/Submit ver/nat.c
@@ -112,7 +112,7 @@ double get_time(){ // in millisecond
 
 int consume_token(){
   curr_time = get_time();
-  int time_diff = ((curr_time - prev_time) % 1000);
+  int time_diff = ((int)(curr_time - prev_time) % 1000);
   printf("fime diff: %d\n", time_diff);
   if(time_diff){
     num_token += fill_rate * time_diff;
